backend/requirements.txt
Flask==3.1.1
flask-cors==6.0.1
tensorflow==2.19.0
numpy==2.1.3
Pillow==11.2.1